backend/requirements.dev.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
joblib==1.3.2
numpy==1.26.1
python-dotenv==1.0.0
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.3
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.8.0
